Check for exception pdus in send_message
For both RTU and TCP client cases. The exception ADU is only 2 bytes
(shorter than all the other ADUs), so we can read those bytes first,
check that there is no error, recv the rest of the expected ADU size,
and concatenate the two for further parsing into the correct response.