Change factory methods to class methods (#103)
Instead of declaring the create_from_request_pdu and create_from_response_pdu factory methods as @staticmethod, use @classmethod. Then the class is passed to the method as the cls parameter, which can be used to instantiate an instance, rather than hardcoding the class name.

This allows the function classes to be subclassed without having to reimplement the factory methods just to change the class name.

Also fixes up some comments that were incorrect.