Use non privileged tcp port in examples (#109)
The server example had port 502 hard coded. This port is a privileged port.
Trying to use it without permissions could result in a failure of the example.
Port 502 is still the default, but a nice error message is printed to warn the user
about the problem. Also a flag has been introduced to change the port.

To keep the example client working with the example server code a similar
change has been made to the example client.