Fix string representation of UndefinedModbusError.
The __repr__() special method is usually called in place of __str__(),
but only if the latter is not defined.  However the BaseException
class already has a __str__() method, so we need to override that
specifically.  Note that this change doesn't touch the other cases
below where the same problem exists.

Initialize the error_code field explicitly using the KeyError argument
instead of passing in the whole exception object.